bug solving

+++ lib/data/model/res_model/get_order_by_id/get_order_by_id_model.dart
@@ -148,7 +148,7 @@ class Product with _$Product {
     @JsonKey(name: "discountedPrice")
     double? discountedPrice,
     @JsonKey(name: "itemWeight")
-    int? itemWeight,
+    double? itemWeight,
     @JsonKey(name: "issueStatus")
     Status? issueStatus,
     @JsonKey(name: "isIssue")
